repocitory mongo add

+++ src/main/java/com/reto/pedidos/repository/PedidoRepository.java
@@ -3,6 +3,6 @@ package com.reto.pedidos.repository;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import com.reto.pedidos.model.Pedido;
 
-public interface PedidoRepository extends ReactiveCrudRepository <Pedido, Long>{
+public interface PedidoRepository extends ReactiveCrudRepository <Pedido, Integer>{
     
 }
